Separate DoubleLineImageCellView into LeftImageCellView and DoubleLineCellView
