+++ BankWallet/BankWallet/UserInterface/Cells/Views/LeftImageCellView.swift
@@ -1,0 +1,27 @@
+import UIKit
+import SnapKit
+
+class LeftImageCellView: UIView {
+    private let imageView = UIImageView()
+
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+
+        addSubview(imageView)
+        imageView.setContentHuggingPriority(.required, for: .horizontal)
+        imageView.snp.makeConstraints { maker in
+            maker.centerY.equalToSuperview()
+            maker.leading.equalToSuperview().offset(AppTheme.margin4x)
+            maker.trailing.equalToSuperview()
+        }
+    }
+
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    func bind(image: UIImage?) {
+        imageView.image = image
+    }
+
+}
